Reload clock and activity report on remote timer changes.

+++ app/controllers/timer_controller.rb
@@ -1,12 +1,23 @@
 class TimerController < ApplicationController
-
   # TODO: https://api.rubyonrails.org/v6.0.1/classes/ActionView/Helpers/FormTagHelper.html#method-i-image_submit_tag
 
   # timer_path: GET /timer
   def index
-    @timer = current_user.timer || Timer.new
-    @activities = get_activities
-    @project_options = project_options
+    respond_to do |format|
+      format.html do
+        # Display the timer page.
+        @timer = current_user.timer || Timer.new
+        @activities = get_activities
+        @project_options = project_options
+      end
+
+      format.json do
+        # Fetch the current timer object as json.
+        timer = current_user.timer
+        #sleep 20 - test dropped response
+        render json: timer
+      end
+    end
   end
 
   # timer_path: POST /timer
@@ -29,50 +40,59 @@ class TimerController < ApplicationController
     @timer.name = name
     @timer.save!
 
-    @project_options = project_options
-    respond_to do |format|
-      format.js { render 'timer_clock' }
-    end
+    replace_clock
   end
 
-  # timer_path: PATCH /timer
-# def update
-#   # TODO: check that project.user == current_user
-#   @timer = current_user.timer
-#   @timer.update!(timer_params)
-#
-#   @project_options = project_options
-#   respond_to do |format|
-#     format.js { render 'timer_clock' }
-#   end
-# end
+  def replace_clock
+    @timer = Timer.find_by_user_id(current_user.id) || Timer.new
+    # puts "===== replace_clock #{@timer.inspect}"
+    @project_options = project_options
+
+    respond_to { |format| format.js { render 'replace_clock' } }
+  end
 
   # timer_finish_path: POST /timer/finish
   def finish
     # TODO: check that project.user == current_user
-    timer = current_user.timer
+    return unless timer = current_user.timer
     project = timer.project
-    activity = Activity.new({
-      user: current_user,
-      client: project.client,
-      project: project,
-      start: timer.start,
-      length: Time.now - timer.start,
-      name: timer.name
-    })
+    activity =
+      Activity.new(
+        {
+          user: current_user,
+          client: project.client,
+          project: project,
+          start: timer.start,
+          length: Time.now - timer.start,
+          name: timer.name
+        }
+      )
     activity.save!
     timer.destroy
 
+    # Call replace_clock for fast feedback.
+    # replace_page will get called by the js global_ticker later on.
+    replace_clock
+  end
+
+  # timer_replace_page_path: GET /timer/replace_page
+  def replace_page
     # Redisplay the page with the updated report.
-    @timer = Timer.new
+    @timer = Timer.find_by_id(current_user.id) || Timer.new
     @activities = get_activities
     @project_options = project_options
-    render 'index'
+
+    respond_to do |format|
+      format.js do
+        # puts "%%% replace_page.js"
+        render 'replace_page'
+      end
+    end
   end
 
   # timer_name_path: POST /timer/name
   def name
-    timer = current_user.timer
+    return unless timer = current_user.timer
     timer.name = params[:name] || ''
     timer.save!
   end
@@ -80,37 +100,38 @@ class TimerController < ApplicationController
   # timer_project_path: POST /timer/project
   def project
     # TODO: check that project.user == current_user
-    timer = current_user.timer
+    return unless timer = current_user.timer
     timer.project_id = params[:project_id].to_i
     timer.save!
   end
 
   # timer_path: DELETE /timer
   def destroy
-    current_user.timer.destroy!
-
+    current_user.timer.destroy
     @timer = Timer.new
-    respond_to do |format|
-      format.js { render 'timer_clock' }
-    end
+    @project_options = project_options
+    respond_to { |format| format.js { render 'replace_clock' } }
   end
 
   private
 
   def get_activities
-    current_user.activities
-      .select('activities.*, projects.name as project_name, clients.name as client_name')
+    current_user.activities.select(
+      'activities.*, projects.name as project_name, clients.name as client_name'
+    )
       .where('start >= ?', 10.days.ago.midnight)
       .order(start: :desc)
       .joins(:project, :client)
   end
 
   def project_options
-    opts = current_user.projects
-      .select('projects.id, projects.name as project_name, clients.name as client_name')
-      .order(updated_at: :desc)
-      .joins(:client)
-      .map do |project|
+    opts =
+      current_user.projects.select(
+        'projects.id, projects.name as project_name, clients.name as client_name'
+      )
+        .order(updated_at: :desc)
+        .joins(:client)
+        .map do |project|
         pn = project.project_name
         cn = project.client_name
         pn = pn.empty? ? 'No Project' : pn
